I implemented the Lasso select in Scatter. There are still some kinks.
But basically, this is probably the framework that we'll be using!
The main obstacles to tackle:
rethinking how SelectionHub.select() should work, now that we might use the Lasso to select multiple unrelated structures.
The previous selection methods always assumed you wanted a single structure plus all its children, but now we might want to split these into two use cases:
1. You pick an object and want all its children
2. You pick a handful of unrelated objects and you don't want their children

in which case it might make sense to reimplement (1.) as a special case of (2.) which activates when you pass the select() function a single structure (and then it goes and populates the selection with all the children before passing it along to the plots) or something like this. You see what I mean?

This kind of thing would necessitate changing some of the logic in, e.g. BasicDendrogramViewer._update_lines() which calls DendrogramPlotter.get_lines() which has the code ```structures = structure.descendants + [structure]``` implying that when you plot some lines you always want descendants. I'm not super sure how hard I want to fix that, but if we wanted to have a perfect logical correlation with the scatter plots you circle and the specific branches that get highlighted, we would need to circumvent this "get all descendants" logic.